backend/requirements.txt
Django~=4.2.6
Pillow~=10.1.0
djangorestframework~=3.14.0
django-autoslug~=1.9.9
psycopg2~=2.9.9
django-imagekit~=5.0.0
djoser~=2.2.2
django-rest-passwordreset~=1.3.0
django-cors-headers~=4.3.1